+++ infrastructure/quick-deploy/aws/monitoring/iam.tf
@@ -29,3 +29,33 @@ resource "aws_iam_policy_attachment" "send_logs_from_fluent_bit_to_cloudwatch_at
   policy_arn = aws_iam_policy.send_logs_from_fluent_bit_to_cloudwatch_policy.0.arn
   roles      = var.eks.self_managed_worker_iam_role_names
 }
+
+# Write objects in S3
+data "aws_iam_policy_document" "write_object" {
+  count = (local.s3_enabled ? 1 : 0)
+  statement {
+    sid = "WriteFromS3"
+    actions = [
+      "s3:PutObject"
+    ]
+    effect = "Allow"
+    resources = [
+      "${var.monitoring.s3.arn}/*"
+    ]
+  }
+}
+
+resource "aws_iam_policy" "write_object" {
+  count       = (local.s3_enabled ? 1 : 0)
+  name_prefix = "s3-logs-write-${var.eks.cluster_name}"
+  description = "Policy for allowing read object in S3 logs ${var.eks.cluster_name}"
+  policy      = data.aws_iam_policy_document.write_object[0].json
+  tags        = local.tags
+}
+
+resource "aws_iam_policy_attachment" "write_object" {
+  count      = (local.s3_enabled ? 1 : 0)
+  name       = "s3-logs-write-${var.eks.cluster_name}"
+  policy_arn = aws_iam_policy.write_object[0].arn
+  roles      = var.eks.self_managed_worker_iam_role_names
+}